Rename isDiffMap to isMapEqual
