+++ packages/api/src/providers/ActivitySendStatus/private/isMapEqual.ts
@@ -3,12 +3,16 @@
  *
  * The equality of the value is compared using `Object.is`.
  */
-export default function isDiffMap<K, V>(x: ReadonlyMap<K, V>, y: ReadonlyMap<K, V>): boolean {
+export default function isMapEqual<K, V>(x: ReadonlyMap<K, V>, y: ReadonlyMap<K, V>): boolean {
+  if ((!x && y) || (x && !y)) {
+    return false;
+  }
+
   const xKeys = new Set(x?.keys());
   const yKeys = new Set(y?.keys());
 
   if (xKeys.size !== yKeys.size) {
-    return true;
+    return false;
   }
 
   for (const key of yKeys) {
@@ -16,14 +20,14 @@ export default function isDiffMap<K, V>(x: ReadonlyMap<K, V>, y: ReadonlyMap<K, 
   }
 
   if (xKeys.size !== yKeys.size) {
-    return true;
+    return false;
   }
 
   for (const key of xKeys) {
     if (!Object.is(x.get(key), y.get(key))) {
-      return true;
+      return false;
     }
   }
 
-  return false;
+  return true;
 }
